refactor(recordings): extract consultee recordings logic to service layer
Move complex business logic from consultee recordings API route into
RecordingService.getConsulteeRecordings() for better separation of
concerns and testability.

Changes:
- Add consulteeRecordingInclude and ConsulteeRecordingWithDetails type
- Add RecordingService.getConsulteeRecordings() method
- Simplify API route from ~255 lines to ~100 lines

Addresses Gemini PR feedback (excluding RouteParams Promise comment
which is incorrect for Next.js 15).
